Remove the patch module from base (#267)
Patch is a posix oriented module but it is community supported and not
very popular.  For instance, assemble (which we do not include) is more
frequently used in galaxy roles than patch.  This PR removes the module
and associated action plugin from base but I think it's on the edge of
what we want so we probably should get an opinion from someone besides
me before making a decision.

(Note: if we keep it, we may want to consider adding assemble as well)