Move timezone from base to community (#266)
Although all posix systems we care about have a timezone, it's not the
most essential part of managing a posix system.  Mount, which we do not
include in base, would seem to be more important, for instance.  It
seems like it is on the edge, though, so someone else besides me should
probably also weigh in so we can have several opinions.